Deleting unused prop "rowsMax"
Deleting unused prop "rowsMax" that is causing the following warning:

react-dom.development.js:67 Warning: React does not recognize the `rowsMax` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `rowsmax` instead. If you accidentally passed it from a parent component, remove it from the DOM element.